Split database models into separate file for each model.
